requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.1.0
requests>=2.31.0
cryptography>=41.0.0

# Email APIs
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
exchangelib>=5.0.0

# PDF Processing (Local)
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
pymupdf>=1.23.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
psycopg2-binary>=2.9.0  # For PostgreSQL (AWS)

# AWS SDK (for production)
boto3>=1.34.0
botocore>=1.34.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Logging and Monitoring
structlog>=23.0.0